+++ insteon_mqtt/mqtt/topic/BaseTopic.py
@@ -3,7 +3,7 @@
 # MQTT Base Topic
 #
 #===========================================================================
-from .. import log
+from ... import log
 
 LOG = log.get_logger()
